fix #42 Rework NettyContext pipeline API
This commit reworks the NettyContext pipeline API, dismissing the
addHandler method and redefining the exact and predictable behavior of
addDecoder.

The new addEncoder method adds handlers on the left hand side of the
pipeline, but after any reactor handler that is marked as "left".

addDecoder adds handlers on the right hand side of the pipeline, but
just before any reactor handler that is marked as "right".

The reactor handlers are marked left or right by a prefix in their name
as defined in NettyPipeline.

This defines a clear sub-section of the pipeline that is managed by the
user, and that standard implementation is mutualized in the new
NettyContextSupport class.