fix #22 Provide a way to replace a decoder/encoder if it already exist
This commit adds the `setDecoder`/`setEncoder` methods to NettyContext,
which will replace a pre-existing handler with the same name.

Additionally, in case of replacement the associated extractor is added
or removed as is relevant.

However, even if the extractor is added or removed the associated
cleanup hooks will still be in place (but they should be idempotent).