Fix custom build version and double quotes
The `versionBranch` version tuning now correctly works if the passed
parameter is enclosed in double quotes, and without double quotes.

It is both acceptable to pass nothin at all (`-PversionBranch=`) or to
pass an "empty String" with double quotes (`-PversionBranch=""`).

Additionally, the version converted to OSGI is printed out to console
when setting the special version.