fix #66 Expose duplicate of options for TCP/HTTP Client and Server
This is useful when the handler cannot be created, to get the original
configuration of the Client/Server object.

The toString() methods for Client/Servers and Options have been aligned,
and each option type has 2 additional String representations (one
detailed as used in toString, the other simplified, as used by the
client/server's toString).