fix #109 Make HttpClient SSL work on relative GET
This commit fixes HttpClient not correctly working with SSL activated
when performing a request (eg. get) on a relative url/path.

This was due to HttpClientOptions.formatSchemeAndHostRelative not taking
the SslContext configuration into account when creating a full URL for
such a request, resulting in HTTP requests being performed on a HTTPS
connection.