issue #89: Update pendingResponses in the event for termination
Problem: pendingResponses is decreased when
r.i.n.h.server.HttpServerHandler#write is invoked with LastHttpContent,
but the channel attributes are cleared in the event for termination
r.i.n.channel.ChannelOperations#onHandlerTerminate.
New request can be received after invocation of the write operation and
before event for termination.
r.i.n.h.server.HttpServerHandler#channelRead
In this situation instead of pipelining the new request the implementation
will check the channel attributes and as they are still not cleared the
request will not be processed.