fix #25 Allow to dispose LoopResources and PoolResources and listen
This commit adds a disposeLater method to both LoopResources and
PoolResources that allow to dispose these but also get a notification of
when the disposal is done, in a best effort fashion.

The call to disposeLater defers actual cleanup until subscription.

TcpResources and HttpResources have an equivalent method for shutdown:
shutdownLater.