fix #80 Add NettyContext config callback, rework Channel config options
This commit adds a configuration callback to NettyOptions that deals
with newly initialized NettyContexts rather than just with the Channel.
This allows to get access to the handler management API for easier
insertion of handlers in the pipelines that have reactor-netty handlers.

Additionally, this commit does a bit of rework and cleanup around other
options that deal with pipeline configuration, notably afterChannelInit,
onChannelInit and channelGroup.