Fix #134: Handle pending writes when outbound completion is reached
In some use cases it may appear that the last flush operation that
is performed by PublisherSender#onComplete/onError will not able
to write the data completely. In these cases the underlaying Netty
implementation will schedule an additional task which will try to
flush the remaining data. This additional flush operation will call
ChannelPipeline#flush. Extend ChannelOperationsHandler#drain so that
it is able to handle these additional callbacks.