fix #146 Make sendFile work with ZipFileSystem, use FileChannel
This commit fixes the case of sending a file from within a zip archive
via a ZipFileSystem. This would fail since #112 due to the use of a
RandomAccessFile instead of a FileChannel. RandomAccessFile needs a File
but the toFile() method is unsupported on a ZipFileSystem

This commit goes back to using FileChannel, and uses ChunkedNioFile
instead of ChunkedFile in the chunked strategy.