fix #144 StartAndAwait alias without the onStarted callback (#149)
The onStarted is useful to get a reference to the facade, in order to
programmatically shut it down later (eg. when running the startAndAwait
from inside a Thread), but is optional. This adds an override without
the callback.

Additionally, do the same for HttpServer.startRouterAndAwait plus add
tests for both startRouter and startRouterAndAwait.