fix #161 Let BlockingNettyContext manage the JVM shutdown hook
This commit lets BlockingNettyContext manage the installation and
removal of its associated JVM shutdown hook. By keeping a reference to
it when installing it, it can be removed when the facade is shut down
manually rather than via the hook, preventing issues at hook execution.

Added installShutdownHook and removeShutdownHook methods to that effect.

The JVM-specific shutdown is isolated in its own private method, which
allows to do specific logs on shutdown mentioning the JVM hook (and a
toString of the hook Thread).