Do not specify SO_LINGER timeout, JVM default will be used
When SO_LINGER >= 0 this will specify a timeout value, in seconds (linger interval).
The linger interval is the timeout for the close method to **block**
while the OS attempts to transmit the unsent data. When **0** the OS will not attempt
to transmit the unsend data.
When SO_LINGER < 0, then the close method does not wait until unsent data is
transmitted, but if possible the OS will transmit any unsent data before
the connection is closed.
More information StandardSocketOptions#SO_LINGER