fix #653 When a channel is acquired and closed immediately re-acquire only once
When a channel is acquired and closed immediately, re-acquire only once.
If the second attempt also fails return an error and let the user decide to retry
or not.
Ensure the channel is released to the pool so that the pool updates the counters,
this is in case of FixedChannelPool.