use WireMock to locally simulate HTTP server
The server is exposed as HttpEchoTestingServer utility class, which
is a JUnit ClassRule and can also be used standalone (has start and
stop methods). It exposes the port it runs on, as well as the base
HTTP url.

Other approaches were explored and discarded:
 - httpin.org in a Docker container (via TestContainers): a bit
 overkill on resources, and not all dev / ci machine will have Docker
 - MockServer: uses Netty AND doesn't shade the dependency, resulting
 in potential Netty version conflics.

WireMock doesn't use Netty dependency like MockServer, so no clash.
Furthermore, there is a shaded standalone version of WireMock.

This commit also polishes some tests to run on local HTTP server:
  - polish tests that don't BAD_REQUEST anymore due to GET having a body
  (not enforced on MockServer, contrary to example.com and google.com)
 - polish one TcpClientTests to also avoid HTTP remote, using the
 HttpEchoTestingServer standalone.