fix #913 Apply BND convention to shadowJar as well
The bnd plugin only applies to the default `jar` task. Since we're now
using the shadow plugin, `shadowJar` has no OSGI data in its MANIFEST.

This commit fixes the situation by applying the bnd plugin convention to
the shadowJar task. It also copies the bnd configuration since this is
not inherited.

The manifest content is tested for key OSGI and Java metadata (not
including the detail of the import/export, though).