Adds failing test for unexpected timing in `HttpClient.doOnResponse()` (#1033)
For metrics and timing, we should be able to get a hook at headers time.
`HttpClient.doOnResponse()` docs say this is the hook for that, except
it clarifies as *after*, not at the moment of the headers time. Small
discrepencies are ok, but right now, the implementation runs the hook
after the `Mono.block()` completes. This makes tracing look strange as
a blocking client should never finish after the next span starts.

There are a couple ways to deal with this, but perhaps only state hooks
need adjusted, as this only breaks consistently on a redirect follow-up
request. Other tests seem to pass consistently.