Expose reactor-netty artifact
This artifact will publish an empty jar with dependencies to
reactor-netty-core and reactor-netty-http, aggregated javadoc and the reference
documentation.
It is not a fat jar as there might be claspath problems when this artifact and
reactor-netty-core/reactor-netty-http are present
It uses jar and not pom packaging in order to provide a migration path.