fix #1406 When connection is closed abnormally, emit WebSocketCloseStatus.ABNORMAL_CLOSURE
According to the specification (below) the emitted WebSocketCloseStatus will be
WebSocketCloseStatus.ABNORMAL_CLOSURE, but the CloseWebSocketFrame
will be still created with status code "-1" as 1006 cannot be set as a status code for
CloseWebSocketFrame.

"1006 is a reserved value and MUST NOT be set as a status code in a
Close control frame by an endpoint.  It is designated for use in
applications expecting a status code to indicate that the
connection was closed abnormally, e.g., without sending or
receiving a Close control frame."