fix #1485 When CloseWebSocketFrame with -1 status code (undefined/empty), emit WebSocketCloseStatus.EMPTY
According to the specification the endpoint may echo the received Close frame,
so Reactor Netty will echo the received Close frame:

https://tools.ietf.org/html/rfc6455#section-5.5.1
"If an endpoint receives a Close frame and did not previously send a
Close frame, the endpoint MUST send a Close frame in response.  (When
sending a Close frame in response, the endpoint typically echos the
status code it received.)"

However for the emitted Close status (the one that is used by the application),
if a status code is not defined or empty, the Close status should be considered 1005,
so Reactor Netty will emit 1005 with WebsocketInbound#receiveCloseStatus.

https://tools.ietf.org/html/rfc6455#section-7.1.5
"If this Close control frame contains no status code, _The WebSocket
Connection Close Code_ is considered to be 1005."