Add HttpDecoderSpec#allowDuplicateContentLengths(boolean)
API for configuring whether or not to allow duplicate Content-Length headers with the same value.
By default this is not allowed and such messages are rejected.
When this is configured to allow duplicate Content-Length headers, they have to have the same value,
otherwise such messages are rejected. The duplicated Content-Length headers are replaced
with a single valid Content-Length header.

Fixes #1636