Expose an API for obtaining all cookies with one and the same name
https://datatracker.ietf.org/doc/html/rfc6265#section-4.2.2

Although cookies are serialized linearly in the Cookie header,
servers SHOULD NOT rely upon the serialization order.  In particular,
if the Cookie header contains two cookies with the same name (e.g.,
that were set with different Path or Domain attributes), servers
SHOULD NOT rely upon the order in which these cookies appear in the
header.

Fixes #1641