Minimum coverage of channel init issue without micrometer
This commit attempts at covering the channel initialization issue when
metrics recorders are registered during configuration but Micrometer is
not on the classpath.

The NoClassDefFoundError itself is not accessible. It is swallowed by
Netty during channel initialization.

This is not ideal but we can still detect something is off, since the
HttpClient does throw a StacklessClosedChannelException (tested on tag
v1.0.11).

See #1847.
See #1848.