New HTTP/2 connection pool implementation
The connection is removed from the pool when:
- The connection is closed.
- The connection has reached its life time and there are no active streams.
- The connection has no active streams.
- When the client is in one of the two modes: 1) H2 and HTTP/1.1 or 2) H2C and HTTP/1.1,
and the negotiated protocol is HTTP/1.1.

The connection is filtered out when:
- The connection has reached its life time and there are active streams. In this case, the connection stays
in the pool, but it is not used. Once there are no active streams, the connection is removed from the pool.
- The connection has reached its max active streams configuration. In this case, the connection stays
in the pool, but it is not used. Once the number of the active streams is below max active streams configuration,
the connection can be used again.

This pool always invalidate the PooledRef, there is no release functionality.
- PoolMetrics#acquiredSize() and PoolMetrics#allocatedSize() always return the number of
the active streams from all connections currently in the pool.
- PoolMetrics#idleSize() always returns 0.

Fixes #1789, #1818, #1844