Run spotless/japicmp as a preliminary job in ci (#2023)
This commit changes the `check_transport.yml` workflow so that it runs a
first job to check spotless and japicmp. The test matrix, which run longer,
depend on that first job succeeding.

We output a few tips to fix spotless failures and to indicate japicmp
failures need to be discussed with the maintainers, as pointers intended
for contributors.

This commit also fixes `japicmp` task in http-brave module to trigger
with the `check` task and to require `downloadBaseline` beforehand.