Add reflection hints for better native-image support
When a `ChannelHandler` is added to the `Netty` channel pipeline and if the `ChannelHandler`
implements `ChannelInboundHandlers`/`ChannelOutboundHandlers`,
`io.netty.channel.ChannelHandlerMask#mask0` will inspect the `ChannelHandler's` methods for
`io.netty.channel.ChannelHandlerMask.Skip` annotation.
When there are no reflection hints, `NoSuchMethodException` is thrown when trying to query the methods with reflection.
This will lead to methods marked as not skippable, although the program will be able to run
without any problems, when all methods are marked as not skippable, one might observe performance issues.

Fixes #2419