Get SslHandler by name for ALPN in HttpClientChannelInitializer (#2488)
Switch to accessing the SslHandler by name in client ALPN.
This should always return the SslHandler for remote endpoint.
If we don't do this, the pipeline will always return the first SslHandler in the pipeline.
When talking through a TLS enabled forward proxy there may be more than one SslHandler present in the pipeline.
This change ensures that we always get the SslHandler for the remote HTTP
endpoint and not an intermediary, ensuring we set the correct negotiated protocol.

Fixes #2480