Ensure HttpServer minCompressionSize is respected when Content-Length is not explicit and response is Mono
When a Mono is sent as a response or a single ByteBuf is sent as a response,
Reactor Netty will send FullHttpResponse and thus is able to set Content-Length if such is not already set.
Based on the Content-Length, Reactor Netty can decide whether or not a compression is needed.