Improving context-propagation tests
Recent change in context-propagation library introduced the ability
to reset ThreadLocal values not present in the sourceContext.

The `ContextSnapshot#captureFrom()` combined with
`ContextSnapshot#setThreadLocals()` removes all ThreadLocal values
not present in the sourceContext.

The method `ContextSnapshot.setAllThreadLocalsFrom()` allows to
only set the values from the provided context, avoiding the need
for a temporary `ContextSnapshot` copy of the values.

Currently, reactor-netty can't upgrade, as 1.0.1 of context-propagation
disallows selective setting of only the interesting values to ThreadLocals
without the need to look up existing ThreadLocals.

However, that will be fixed and the `setAllThreadLocalsFrom`
is a proper shortcut going forward.