Wrong host port 80 when only X-Forwarded-Proto: https is present (#2773)
When a front-end proxy forwards an https request to the backend application in plain http, and if the forwarded request only contain the original Host header without any port (443 by default) as well as "X-Forwarded-Proto: https" only, then the wrong default 80 port is computed by the reactor netty default forwarded header handler. The expected default port should be 443 instead of 80.