Avoid null metric recorder remoteAddress parameters (#2755)
When a reactor-netty http server receives a request with an invalid request URI, metric recorders are invoked with null host/remote addresses. When the URL can’t be parsed, HttpServerOperations.sendDecodingFailures sends a 400 bad request with a FailedHttpServerRequest that is using a null connectionInfo parameter. Moreover, recorders are also called with null host/remote addresses when Unix Domain Sockets are used. 

Modification:

The ConnectionInfo object is now used when sending a decoding error. Now, if the ConnectionInfo could not be created in some rare conditions like when the Host or Forwarded/X-Forwarded headers are malformed, then we fallback on established connection info. Also added support for Unix Domain Sockets.

Fixes #2748