Ensure DefaultHttp2GoAwayFrame is released (#2835)
Rationale: The GOAWAY undergoes a retainedDuplicate() in Http2MultiplexHandler, but is never released. This causes a leak of the underlying packet data (or SSL decoded buffer if SSL is used)

The solution is to make sure the user events are released in ChannelOperationsHandler. The DefaultChannelPipeline does this normally, but ChannelOperationsHandler's consumes the user events without releasing them.