Ensure DisposedChannelConfig#setAutoRead does not change auto-read configuration (#3581)
DisposedChannel is effective when request/response is terminated and replaces the actual channel.
At that point DisposedChannelConfig#setAutoRead must be non operational
as the inbound has already been read and the actual channel has already set auto-read to true.

The issue is also observed with the reproducible example provided by #3495

Fixes #3559