Revert changes in check_netty_4_1_snapshots.yml
`check_netty_4_1_snapshots.yml` workflow is executed against Reactor Netty version `1.2.x`. In this version there is only integration with `netty-incubator-transport-native-io_uring` as this version is built against `Netty 4.1`.
The build against `Netty 4.2` is just to verify Netty backwards compatibility and not to test the new features coming with `Netty 4.2`.

Signed-off-by: Violeta Georgieva <696661+violetagg@users.noreply.github.com>